Parsing Discord Client Instance to Routes
