+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "zander-web",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "description": "",
   "main": "app.js",
   "scripts": {
